testing using actions to update registry instead
